SOLR-16961: Remove deprecated CircuitBreakerManager in 10.0
