+++ solr/core/src/java/org/apache/solr/handler/ContentStreamHandlerBase.java
@@ -139,8 +139,7 @@ public abstract class ContentStreamHandlerBase extends RequestHandlerBase {
         rsp.add(STATUS, FAILURE);
         rsp.setException(
             new SolrException(
-                CircuitBreaker.getErrorCode(trippedCircuitBreakers),
-                "Circuit Breakers tripped " + errorMessage));
+                CircuitBreaker.ERROR_CODE, "Circuit Breakers tripped " + errorMessage));
         return true;
       }
     }
